✨ feat: update structure repository of users
Add Singleton Pattern at InMemoryRepository
